--- app/src/main/java/de/info3/wegfinder24/newtwork/JSON/Example.java
@@ -1,56 +1,0 @@
-package de.info3.wegfinder24.newtwork.JSON;
-
-import java.util.List;
-import javax.annotation.Generated;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("jsonschema2pojo")
-public class Example {
-
-    @SerializedName("type")
-    @Expose
-    private String type;
-    @SerializedName("features")
-    @Expose
-    private List<Feature> features = null;
-    @SerializedName("bbox")
-    @Expose
-    private List<Double> bbox = null;
-    @SerializedName("metadata")
-    @Expose
-    private Metadata metadata;
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public List<Feature> getFeatures() {
-        return features;
-    }
-
-    public void setFeatures(List<Feature> features) {
-        this.features = features;
-    }
-
-    public List<Double> getBbox() {
-        return bbox;
-    }
-
-    public void setBbox(List<Double> bbox) {
-        this.bbox = bbox;
-    }
-
-    public Metadata getMetadata() {
-        return metadata;
-    }
-
-    public void setMetadata(Metadata metadata) {
-        this.metadata = metadata;
-    }
-
-}
