+++ app/src/main/java/de/info3/wegfinder24/newtwork/Segment.java
@@ -1,11 +1,12 @@
-package de.info3.wegfinder24.newtwork.JSON;
+package de.info3.wegfinder24.newtwork;
 
+import java.util.List;
 import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 @Generated("jsonschema2pojo")
-public class Summary {
+public class Segment {
 
     @SerializedName("distance")
     @Expose
@@ -13,6 +14,9 @@ public class Summary {
     @SerializedName("duration")
     @Expose
     private Double duration;
+    @SerializedName("steps")
+    @Expose
+    private List<Step> steps = null;
 
     public Double getDistance() {
         return distance;
@@ -28,6 +32,14 @@ public class Summary {
 
     public void setDuration(Double duration) {
         this.duration = duration;
+    }
+
+    public List<Step> getSteps() {
+        return steps;
+    }
+
+    public void setSteps(List<Step> steps) {
+        this.steps = steps;
     }
 
 }
