+++ app/src/main/java/de/info3/wegfinder24/newtwork/Step.java
@@ -1,4 +1,4 @@
-package de.info3.wegfinder24.newtwork.JSON;
+package de.info3.wegfinder24.newtwork;
 
 import java.util.List;
 import javax.annotation.Generated;
