+++ app/src/main/java/de/info3/wegfinder24/newtwork/Summary.java
@@ -1,12 +1,11 @@
-package de.info3.wegfinder24.newtwork.JSON;
+package de.info3.wegfinder24.newtwork;
 
-import java.util.List;
 import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 @Generated("jsonschema2pojo")
-public class Segment {
+public class Summary {
 
     @SerializedName("distance")
     @Expose
@@ -14,9 +13,6 @@ public class Segment {
     @SerializedName("duration")
     @Expose
     private Double duration;
-    @SerializedName("steps")
-    @Expose
-    private List<Step> steps = null;
 
     public Double getDistance() {
         return distance;
@@ -32,14 +28,6 @@ public class Segment {
 
     public void setDuration(Double duration) {
         this.duration = duration;
-    }
-
-    public List<Step> getSteps() {
-        return steps;
-    }
-
-    public void setSteps(List<Step> steps) {
-        this.steps = steps;
     }
 
 }
